Move some CDK stuff around
